requirements.txt
requests
ics
gitpython
